Plugsity Complete

+++ src/controllers/address.controllers.js
@@ -11,7 +11,7 @@ class AddressController {
   //@access public
   getAllAddresses = async (req, res) => {
     Logger.info(`Request received: ${req.method} ${req.url}`);
-  
+
     const addresses = await AddressService.find();
 
     if (!addresses) {
@@ -25,7 +25,7 @@ class AddressController {
       .body({ addresses })
       .send();
   };
-  
+
   //@desc get address
   //@route GET /api/address/:addressId
   //@access public
@@ -34,17 +34,13 @@ class AddressController {
 
     const { addressId } = req.params;
     const address = await AddressService.findById(addressId);
-      
+
     if (!address) {
       throw new HttpError(404, "Address not found");
     }
 
     Logger.info(`Address: ${address}`);
-    Response(res)
-      .status(200)
-      .message("Address")
-      .body({ address })
-      .send();
+    Response(res).status(200).message("Address").body({ address }).send();
   };
 
   //@desc create address by userId
@@ -57,14 +53,18 @@ class AddressController {
     const addressData = req.body;
 
     const user = await UserService.findById(userId);
-    if(!user){
+    if (!user) {
       Logger.error(`User not found`);
-      throw new HttpError(404, "User not found")
+      throw new HttpError(404, "User not found");
+    }
+
+    if (userId.toString() !== req.user._id.toString()) {
+      throw new HttpError(404, "Unauthorized!");
     }
 
     const address = await AddressService.create({ userId, ...addressData });
 
-    if(address){
+    if (address) {
       Logger.info(`Address created: ${address}`);
       Response(res)
         .status(201)
@@ -84,14 +84,17 @@ class AddressController {
     Logger.info(`Request received: ${req.method} ${req.url}`);
 
     const { addressId } = req.params;
+
+    const address = await AddressService.findById(addressId);
+
+    if (address.userId.toString() !== req.user._id.toString()) {
+      throw new HttpError(404, "Unauthorized!");
+    }
+
     await AddressService.findByIdAndUpdate(addressId, { isDefault: true });
 
     Logger.info(`Default address set`);
-    Response(res)
-      .status(200)
-      .message("Default address set")
-      .body()
-      .send();
+    Response(res).status(200).message("Default address set").body().send();
   };
 
   //@desc update address
@@ -103,7 +106,18 @@ class AddressController {
     const { addressId } = req.params;
     const addressData = req.body;
 
-    const updatedAddress = await AddressService.findByIdAndUpdate(addressId, addressData);
+    const address = await AddressService.findById(addressId);
+
+    if (
+      address.userId.toString() !== req.user._id.toString() &&
+      req.user.role.toString() !== "Admin"
+    ) {
+      throw new HttpError(404, "Unauthorized!");
+    }
+
+    Object.assign(address, addressData);
+
+    const updatedAddress = await address.save();
 
     Logger.info(`Address updated: ${updatedAddress}`);
     Response(res)
@@ -120,16 +134,21 @@ class AddressController {
     Logger.info(`Request received: ${req.method} ${req.url}`);
 
     const { addressId } = req.params;
+
+    const address = await AddressService.findById(addressId);
+
+    if (
+      address.userId.toString() !== req.user._id.toString() &&
+      req.user.role.toString() !== "Admin"
+    ) {
+      throw new HttpError(404, "Unauthorized!");
+    }
+
     await AddressService.findByIdAndDelete(addressId);
 
     Logger.info(`Address deleted`);
-    Response(res)
-      .status(200)
-      .message("Address deleted")
-      .body()
-      .send();
+    Response(res).status(200).message("Address deleted").body().send();
   };
-
 }
 
 module.exports.AddressController = new AddressController();
